+++ test/unit/fuzz-tools.test.js
@@ -496,6 +496,7 @@ import * as childhoodActV518 from '../../lib/childhood-act-v518.js';
 import * as eckardtV519 from '../../lib/eckardt-v519.js';
 import * as spigelmanV520 from '../../lib/spigelman-v520.js';
 import * as pcPtsd5V521 from '../../lib/pc-ptsd5-v521.js';
+import * as pcdaiV522 from '../../lib/pcdai-v522.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -937,6 +938,7 @@ const MODULES = {
   'eckardt-v519.js': eckardtV519,
   'spigelman-v520.js': spigelmanV520,
   'pc-ptsd5-v521.js': pcPtsd5V521,
+  'pcdai-v522.js': pcdaiV522,
   'tb-testing.js': tbTesting,
 };